+++ atlasdb-dbkvs/src/main/java/com/palantir/atlasdb/keyvalue/dbkvs/DdlConfig.java
@@ -17,6 +17,7 @@ package com.palantir.atlasdb.keyvalue.dbkvs;
 
 import org.immutables.value.Value;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.google.common.base.Preconditions;
@@ -59,6 +60,12 @@ public abstract class DdlConfig {
     @Value.Default
     public int mutationBatchSizeBytes() {
         return 2 * 1024 * 1024;
+    }
+
+    @Value.Default
+    @JsonIgnore
+    public boolean useTableMapping() {
+        return true;
     }
 
     @Value.Check